FIX: jump targets inside a topic list when using priority groups
On a topic list, the post number target makes sense to navigate to the priority
group's post directly.

However, this causes an issue when inside the topic as it will appear that
all tracked posts prior to the post are gone (IE: lower priority posts made before).

Update functionality to have two jump targets: one from outside the topic
to the priority group's post, and another when jumping to the first tracked post
within a topic.